#54: Refactor RootPage, db class
